Remove FileStreamBufferPrivate
