+++ projects/AFile/FileStreamBuffer.cpp
@@ -1,83 +1,65 @@
-// Copyright 2020 Michael Vlach
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
 #include "pch.hpp"
 
 #include "FileStreamBuffer.hpp"
-#include "FileStreamBufferPrivate.hpp"
 
 namespace afile
 {
 FileStreamBuffer::FileStreamBuffer(const char *filename) :
-    d(new FileStreamBufferPrivate(filename))
-{
-}
+    mName(filename),
+    mFile(filename, openMode(filename)),
+    mSize(static_cast<acore::size_type>(std::filesystem::file_size(filename)))
 
-FileStreamBuffer::FileStreamBuffer(FileStreamBuffer &&other) noexcept
 {
-    *this = std::move(other);
 }
 
 FileStreamBuffer::~FileStreamBuffer() noexcept
 {
-    delete d;
+    if (mFile.is_open())
+    {
+        mFile.close();
+        std::filesystem::resize_file(mName, static_cast<size_t>(mSize));
+    }
 }
 
 const char *FileStreamBuffer::filename() const noexcept
 {
-    return d->filename();
+    return mName.c_str();
 }
 
 void FileStreamBuffer::flush()
 {
-    d->flush();
+    mFile.flush();
 }
 
 bool FileStreamBuffer::isOpen() const
 {
-    return d->isOpen();
+    return mFile.is_open();
 }
 
 void FileStreamBuffer::read(acore::size_type index, char *buffer, acore::size_type count)
 {
-    d->read(index, buffer, count);
-}
-
-void FileStreamBuffer::resize(acore::size_type size)
-{
-    d->resize(size);
-}
-
-acore::size_type FileStreamBuffer::size() const noexcept
-{
-    return d->size();
+    mFile.seekg(index);
+    mFile.read(buffer, count);
 }
 
 void FileStreamBuffer::write(acore::size_type index, const char *buffer, acore::size_type count)
 {
-    d->write(index, buffer, count);
+    mFile.seekp(index);
+    mFile.write(buffer, count);
 }
 
-FileStreamBuffer &FileStreamBuffer::operator=(FileStreamBuffer &&other) noexcept
+std::ios_base::openmode FileStreamBuffer::openMode(const std::string &filename) const
 {
-    if (this != &other)
+    if (std::filesystem::exists(filename))
     {
-        delete d;
-        d = other.d;
-        other.d = nullptr;
+        return static_cast<std::ios_base::openmode>(OpenMode::Binary | OpenMode::In | OpenMode::Out | OpenMode::AtEnd);
     }
 
-    return *this;
+    return static_cast<std::ios_base::openmode>(OpenMode::Binary | OpenMode::In | OpenMode::Out | OpenMode::Truncate);
+}
+
+FileStreamBuffer::OpenMode operator|(FileStreamBuffer::OpenMode left, FileStreamBuffer::OpenMode right) noexcept
+{
+    return static_cast<FileStreamBuffer::OpenMode>(static_cast<std::underlying_type<FileStreamBuffer::OpenMode>::type>(left) | static_cast<std::underlying_type<FileStreamBuffer::OpenMode>::type>(right));
 }
 }
